moved role name logic to common
